requirements.txt
fastapi
uvicorn
requests
uvloop
aiohttp
redis
itsdangerous
jinja2
python-dotenv
httpx
sqlalchemy
psycopg2


aioredis
motor
websockets
mongodb-community
